pyproject.toml
[tool.poetry]
name = "dimatech"
version = "0.1.0"
description = ""
authors = ["Lanterman <splen1011@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.35.0"
dotenv = "^0.9.9"
pyjwt = "^2.10.1"
alembic = "^1.16.4"
asyncpg = "^0.30.0"
fastapi = {extras = ["standart"], version = "^0.116.1"}
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.43"
pydantic = {extras = ["email"], version = "^2.11.7"}
black = "^25.1.0"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
